secrets: rename valgrind_mem_requests to mem_requests
In the future, we might also want to use memory sanitizer instrumentation.
